% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{coef.spar}
\alias{coef.spar}
\title{coef.spar}
\usage{
\method{coef}{spar}(spar_res, nummod = NULL, lambda = NULL)
}
\arguments{
\item{spar_res}{result of spar function of class "spar".}

\item{nummod}{number of models used to form coefficients; value with minimal validation Dev is used if not provided.}

\item{lambda}{threshold level used to form coefficients; value with minimal validation Dev is used if not provided.}
}
\value{
List of coefficients with elements
\itemize{
\item intercept
\item beta
\item nummod
\item lambda
}
}
\description{
Extract coefficients from spar object
}
