% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{constructor_screeningcoef}
\alias{constructor_screeningcoef}
\title{Constructor function for building screeningcoef objects}
\usage{
constructor_screeningcoef(name, generate_scrcoef)
}
\arguments{
\item{name}{character}

\item{generate_scrcoef}{function for generating the screening coefficient. This
function should have with arguments \code{scrcoef}, which is a screeningcoef
object, and \code{data}, which is a list of x (matrix of predictors used as input in
\link{\code{spar}}) and y (vector of responses used in  \link{\code{spar}}).}
}
\value{
a function which in turn creates a function which in turn creates an
}
\description{
The created function will return a object of class \code{"screeningcoef"} which
constitutes of a list. The attributes of the generating object will include by
default \code{type}, which can take one of two values \code{"prob"} (indicating
probabilistic screening should be employed) or
\code{"fixed"} (indicating that the top \code{nscreen} variables should be employed).
}
\details{
Creates an object class screeningcoef using arguments passed by user.
}
