% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R
\name{generate_cw}
\alias{generate_cw}
\title{Sparse embedding matrix}
\usage{
generate_cw(rp, m, included_vector)
}
\arguments{
\item{m}{goal dimension, which will be randomly sampled in the SPAR algorithm}

\item{included_vector}{integer vector of column indices for the variables to be
included in the random projection. These indices are produced in the
screening step of the SPAR algorithm.}
}
\value{
(possibly sparse) matrix with m rows and
\code{length(included_vector)} columns.
}
\description{
Sparse embedding matrix
}
\keyword{internal}
