% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_RPM.R
\name{generate_cw_rp}
\alias{generate_cw_rp}
\title{Generate CW Random Projection Matrix (Definition 2.2. in Parzer, Vana_Gür, Filzmoser 2023)}
\usage{
generate_cw_rp(type.rpm, m, p, coef = NULL, ...)
}
\arguments{
\item{m}{goal dimension}

\item{p}{initial dimension}

\item{coef}{p-vector of diagonal elements, random +/-1 by default}
}
\value{
sparse mxp matrix of class "dgCMatrix" (Matrix package)
}
\description{
Generate CW Random Projection Matrix (Definition 2.2. in Parzer, Vana_Gür, Filzmoser 2023)
}
\keyword{internal}
