% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_RPM.R
\name{generate_gaussian_rp}
\alias{generate_gaussian_rp}
\title{Gaussian Random Projection Matrix}
\usage{
generate_gaussian_rp(m, p)
}
\arguments{
\item{m}{goal dimension}

\item{p}{dimension of predictors to be projected}
}
\value{
an m x p matrix
}
\description{
Gaussian Random Projection Matrix
}
\examples{
generate_gaussian_rp(10, 20)
}
