% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{generate_scrcoef_corr}
\alias{generate_scrcoef_corr}
\title{Screening coefficient based on correlation}
\usage{
generate_scrcoef_corr(scrcoef, data)
}
\description{
Screening coefficient based on correlation
}
