% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{generate_scrcoef_glmnet}
\alias{generate_scrcoef_glmnet}
\title{Screening coefficient based  on glmnet coefficients}
\usage{
generate_scrcoef_glmnet(object, data)
}
\description{
Screening coefficient based  on glmnet coefficients
}
