% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{generate_scrcoef_marglik}
\alias{generate_scrcoef_marglik}
\title{Screening coefficient based on marginal GLMs}
\usage{
generate_scrcoef_marglik(scrcoef, data)
}
\description{
Screening coefficient based on marginal GLMs
}
