% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{generate_scrcoef_ridge}
\alias{generate_scrcoef_ridge}
\title{Screening coefficient based  on ridge coefficients}
\usage{
generate_scrcoef_ridge(scrcoef, data)
}
\description{
Screening coefficient based  on ridge coefficients
}
