% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_RPM.R
\name{generate_sparse_rp}
\alias{generate_sparse_rp}
\title{Sparse Random Projection Matrix}
\usage{
generate_sparse_rp(m, p, psi = NULL)
}
\arguments{
\item{m}{goal dimension}

\item{p}{dimension of predictors to be projected}

\item{psi}{a vector of length one representing the parameter of the random projection 0 < vals <= 1}
}
\value{
an m x p matrix of class "dgCMatrix"
}
\description{
Sparse Random Projection Matrix
}
\examples{
generate_sparse_rp(10, 20)
}
