% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R
\name{get_rp}
\alias{get_rp}
\title{Function which works on all random projection objects}
\usage{
get_rp(rp, m, included_vector)
}
\arguments{
\item{rp}{an object of class randomprojection}

\item{m}{integer goal dimension of the projection}

\item{included_vector}{a vector containing column index of the original variables in the
\code{x} matrix to be projected using the random projection}
}
\description{
Function which works on all random projection objects
}
\keyword{internal}
