% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{get_screencoef}
\alias{get_screencoef}
\title{Function which generates the screening coefficients}
\usage{
get_screencoef(y, x, object)
}
\arguments{
\item{y}{vector of responses}

\item{x}{matrix of predictors}

\item{object}{an object of class screencoef}
}
\description{
Function which generates the screening coefficients
}
\keyword{internal}
