% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{plot.spar}
\alias{plot.spar}
\title{plot.spar}
\usage{
\method{plot}{spar}(
  spar_res,
  plot_type = c("Dev", "numAct", "res-vs-fitted"),
  plot_along = c("lambda", "nummod"),
  nummod = NULL,
  lambda = NULL,
  xfit = NULL,
  yfit = NULL
)
}
\arguments{
\item{spar_res}{result of spar function of class "spar".}

\item{plot_type}{one of c("Dev","numAct","res-vs-fitted").}

\item{plot_along}{one of c("lambda","nummod"); ignored when plot_type="res-vs-fitted".}

\item{nummod}{fixed value for nummod when plot_along="lambda" for plot_type="Dev" or "numAct"; same as for predict.spar when plot_type="res-vs-fitted".}

\item{lambda}{fixed value for lambda when plot_along="nummod" for plot_type="Dev" or "numAct"; same as for predict.spar when plot_type="res-vs-fitted".}

\item{xfit}{data used for predictions in "res-vs-fitted".}

\item{yfit}{data used for predictions in "res-vs-fitted".}
}
\value{
ggplot2 object
}
\description{
Plot errors or number of active variables over different thresholds or number of models of spar result, or residuals vs fitted
}
