% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARcv.R
\name{predict.spar.cv}
\alias{predict.spar.cv}
\title{predict.spar.cv}
\usage{
\method{predict}{spar.cv}(
  spar_res,
  xnew,
  opt_par = c("1se", "best"),
  nummod = NULL,
  lambda = NULL,
  coef = NULL
)
}
\arguments{
\item{spar_res}{result of spar function of class "spar".}

\item{xnew}{matrix of new predictor variables; must have same number of columns as x.}

\item{opt_par}{one of c("1se","best"), chooses whether to select the best pair of lambdas and nummods according to CV-MSE, or the sparsest solution within one sd of that optimal CV-MSE;
ignored when nummod and lambda, or coef are given}

\item{nummod}{number of models used to form coefficients; value with minimal validation MSE is used if not provided.}

\item{lambda}{threshold level used to form coefficients; value with minimal validation MSE is used if not provided.}

\item{coef}{optional; result of coef.spar.cv, can be used if coef.spar.cv has already been called.}
}
\value{
Vector of predictions
}
\description{
Predict responses for new predictors from spar object
}
