% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{print.screeningcoef}
\alias{print.screeningcoef}
\title{print.screeningcoef}
\usage{
\method{print}{screeningcoef}(x, ...)
}
\arguments{
\item{x}{description}

\item{...}{further arguments passed to or from other methods}
}
\value{
text summary
}
\description{
Print method for screeningcoef object
}
