% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{print.spar}
\alias{print.spar}
\title{print.spar}
\usage{
\method{print}{spar}(x, ...)
}
\arguments{
\item{x}{result of spar function of class "spar".}

\item{...}{further arguments passed to or from other methods}
}
\value{
text summary
}
\description{
Print summary of spar result
}
