% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{screen_corr}
\alias{screen_corr}
\title{Screening coefficient based  on correlation}
\usage{
screen_corr(..., control = list())
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the
"\code{screeningcoef}" object}

\item{control}{list of controls to be passed to the screening function}
}
\value{
object of class "\code{screeningcoef}" which is a list with elements
\code{name} (character), \code{generate_scrcoef} (function), and \code{control}
(list of controls passes as an argument)
}
\description{
No arguments need to be passed.
}
\details{
Creates an object class "\code{screeningcoef}" using arguments passed by user.
}
