% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{screen_glmnet}
\alias{screen_glmnet}
\title{Screening coefficient based  on glmnet coefficients}
\usage{
screen_glmnet(..., control = list())
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the
"\code{screencoef}" object}

\item{control}{list of controls to be passed to the screening function}
}
\value{
object of class "\code{screencoef}" which is a list with elements

\itemize{
\item \code{name} (character)
\item \code{control} (list of controls passed as an argument)
\item \code{generate_fun}  for generating the screening coefficient. This function should have arguments \code{object}, which is a "\code{screencoef}" object, and \code{data}, which is a list of two elements \code{x} and \code{y} containing the matrix of standardized predictors and the vector of (standardized for Gaussian) responses.
}
}
\description{
Relies on \link[glmnet]{glmnet}.
}
\details{
Creates an object class "\code{screencoef}" using arguments passed by user.
}
