% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{screen_marglik}
\alias{screen_marglik}
\title{Screening coefficient based on marginal GLMs}
\usage{
screen_marglik(...)
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the random
projection matrix}
}
\value{
object of class screeningcoef which is a list with elements name,
generate_scrcoef, control
}
\description{
No arguments need to be passed.
}
\details{
Creates an object class screeningcoef using arguments passed by user.
}
