\name{SPAr-package}
\alias{SPAr-package}
\alias{SPAr}
\docType{package}
\title{
Rare variants association tests based on summation of partition approaches
}
\description{
Perform nonparametric association tests between a set of rare variants and continous or dichotomous outcomes incorporation gene-gene and gene-environmental interactions
}
\details{
\tabular{ll}{
Package: \tab SPAr\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-02-20\cr
License: \tab GPL(>=2)\cr
}
}
\author{
Ruixue Fan

Maintainer: Ruixue Fan <rf2283@columbia.edu>
}
\references{
Fan, R., Lo, S-H (2013). A robust model-free approach for rare variantss association studies incorporating gene-gene and gene-environmental interactions. \emph{PLOS ONE} 8(12):e83057.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
