\name{SPA_example}
\alias{SPA_example}
\docType{data}
\title{
	Example data for SPA with continuous and dichotomous traits and an environmental factor.
}
\description{
	Example data for SPA with continuous and dichotomous traits and an environmental factor.
}
\format{
	SPA_example contains the following objects:
	\describe{
		\item{x}{a numeric genotype matrix of 1000 individuals and 20 SNPs. 
			Each row represents a different individual and each column represents a different SNP marker. }
		\item{y.c}{a numeric vector of continuous phenotypes.}
		\item{y.d}{a numeric vector of dichotomous phenotypes.}
		\item{E}{a numeric vector of an environmental factor.}
	}
}
