% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{spar}
\alias{spar}
\title{Sparse Projected Averaged Regression}
\usage{
spar(
  x,
  y,
  family = gaussian("identity"),
  model = NULL,
  rp = NULL,
  screencoef = NULL,
  xval = NULL,
  yval = NULL,
  nnu = 20,
  nus = NULL,
  nummods = c(20),
  measure = c("deviance", "mse", "mae", "class", "1-auc"),
  parallel = FALSE,
  inds = NULL,
  RPMs = NULL,
  seed = NULL,
  set.seed.iteration = FALSE,
  ...
)
}
\arguments{
\item{x}{n x p numeric matrix of predictor variables.}

\item{y}{quantitative response vector of length n.}

\item{family}{a \link[stats]{family}  object used for the marginal generalized linear model,
default \code{gaussian("identity")}.}

\item{model}{function creating a "\code{sparmodel}" object; defaults to \code{spar_glmnet()}.}

\item{rp}{function creating a "\code{randomprojection}" object.}

\item{screencoef}{function creating a "\code{screeningcoef}" object}

\item{xval}{optional matrix of predictor variables observations used for
validation of threshold nu and number of models; \code{x} is used
if not provided.}

\item{yval}{optional response observations used for validation of
threshold nu and number of models; \code{y} is used if not provided.}

\item{nnu}{number of different threshold values \eqn{\nu} to consider for thresholding;
ignored when nus are given; defaults to 20.}

\item{nus}{optional vector of \eqn{\nu}'s to consider for thresholding;
if not provided, \code{nnu} values ranging from 0 to the maximum absolute
marginal coefficient are used.}

\item{nummods}{vector of numbers of marginal models to consider for
validation; defaults to \code{c(20)}.}

\item{measure}{loss to use for validation; defaults to \code{"deviance"}
available for all families. Other options are \code{"mse"} or \code{"mae"}
(between responses and predicted means, for all families),
\code{"class"} (misclassification error) and
\code{"1-auc"} (one minus area under the ROC curve) both just for
binomial family.}

\item{parallel}{assuming a parallel backend is loaded and available, a logical indicating whether the function should use it. Defaults to FALSE.}

\item{inds}{optional list of index-vectors corresponding to variables kept
after screening in each marginal model of length \code{max(nummods)};
dimensions need to fit those of RPMs.}

\item{RPMs}{optional list of projection matrices used in each
marginal model of length \code{max(nummods)}, diagonal elements will be
overwritten with a coefficient only depending on the given \code{x} and \code{y}.}

\item{seed}{integer seed to be set at the beginning of the SPAR algorithm. Default to NULL, in which case no seed is set.}

\item{set.seed.iteration}{a boolean indicating whether a different seed should be set in each marginal model \code{i}.
This will be set to  \code{seed + i}.}

\item{...}{further arguments mainly to ensure back-compatibility}
}
\value{
object of class \code{"spar"} with elements
\itemize{
\item \code{betas} p x \code{max(nummods)} sparse matrix of class
\code{"\link[=dgCMatrix-class]{dgCMatrix}"} containing the
standardized coefficients from each marginal model
\item \code{intercepts} used in each marginal model
\item \code{scr_coef} p-vector of coefficients used for screening for standardized predictors
\item \code{inds} list of index-vectors corresponding to variables kept after screening in each marginal model of length max(nummods)
\item \code{RPMs} list of projection matrices used in each marginal model of length \code{max(nummods)}
\item \code{val_res} \code{data.frame} with validation results (validation measure
and number of active variables) for each element of \code{nus} and \code{nummods}
\item \code{val_set} logical flag, whether validation data were provided;
if \code{FALSE}, training data were used for validation
\item \code{nus} vector of \eqn{\nu}'s considered for thresholding
\item \code{nummods} vector of numbers of marginal models considered for validation
\item \code{ycenter} empirical mean of initial response vector
\item \code{yscale} empirical standard deviation of initial response vector
\item \code{xcenter} p-vector of empirical means of initial predictor variables
\item \code{xscale} p-vector of empirical standard deviations of initial predictor variables
\item \code{rp} an object of class "\code{randomprojection}"
\item \code{screencoef} an object of class "\code{screeningcoef}"
}
If a parallel backend is registered and \code{parallel = TRUE},
the \link[foreach]{foreach} function
is used to estimate the marginal models in parallel.
}
\description{
Apply Sparse Projected Averaged Regression to High-dimensional Data
\insertCite{parzer2024lm}{spar} \insertCite{parzer2024glms}{spar}.
This function performs the procedure for given thresholds nu and numbers
of marginal models, and acts as a help-function for the full cross-validated
procedure \link{spar.cv}.
}
\examples{
\dontrun{
data("example_data")
spar_res <- spar(example_data$x,example_data$y,
xval=example_data$xtest,yval=example_data$ytest,nummods=c(5,10,15,20,25,30))
spar_res
coefs <- coef(spar_res)
pred <- predict(spar_res,xnew=example_data$x)
plot(spar_res)
plot(spar_res,"Val_Meas","nummod")
plot(spar_res,"Val_numAct","nu")
plot(spar_res,"coefs",prange=c(1,400))}
}
\references{
{
\insertRef{parzer2024lm}{spar}

\insertRef{parzer2024glms}{spar}

\insertRef{Clarkson2013LowRankApprox}{spar}

\insertRef{ACHLIOPTAS2003JL}{spar}
}
}
\seealso{
\link{spar.cv},\link{coef.spar},\link{predict.spar},\link{plot.spar},\link{print.spar}
}
