% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARcv.R
\name{coef.spar.cv}
\alias{coef.spar.cv}
\title{coef.spar.cv}
\usage{
\method{coef}{spar.cv}(spar_res, opt_par = c("best", "1se"), nummod = NULL, lambda = NULL)
}
\arguments{
\item{spar_res}{result of spar.cv function of class "spar.cv".}

\item{opt_par}{one of c("1se","best"), chooses whether to select the best pair of lambdas and nummods according to CV-Dev, or the sparsest solution within one sd of that optimal CV-Dev;
ignored when nummod and lambda are given}

\item{nummod}{optional number of models used to form coefficients}

\item{lambda}{optional threshold level used to form coefficients}
}
\value{
List of coefficients with elements
\itemize{
\item intercept
\item beta
\item nummod
\item lambda
}
}
\description{
Extract coefficients from spar object
}
