% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R
\name{constructor_rp}
\alias{constructor_rp}
\title{Constructor function for building randomprojection objects}
\usage{
constructor_rp(name, generate_fun, update_data_fun = NULL)
}
\arguments{
\item{name}{character}

\item{generate_fun}{function for generating the random projection matrix. This
function should have with arguments \code{rp}, which is a randomprojection
object, \code{m}, the target dimension and a vector of indexes
\code{included_vector} which shows the column index of the original variables in the
\code{x} matrix to be projected using the random projection. This is needed
due to the fact that screening can be employed pre-projection.}

\item{update_data_fun}{function for updating the randomprojection object with
information from the data. This
function should have with arguments \code{rp}, which is a randomprojection
object and \code{data}, which is a list containing \code{x} (the matrix of predictors)
and \code{y} the vector of responses.}
}
\value{
a function which in turn creates an object of class randomprojection
}
\description{
No arguments need to be passed. The entries of the matrix are generated from
a standard normal distribution.
}
\details{
Creates an object class randomprojection using arguments passed by user.
}
