% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{generate_scrcoef_glmnet}
\alias{generate_scrcoef_glmnet}
\title{Screening coefficient based  on glmnet coefficients}
\usage{
generate_scrcoef_glmnet(object, data)
}
\arguments{
\item{object}{"\code{screencoef}" object}

\item{data}{list of x and y}
}
\value{
vector of screening coefficients of length p
}
\description{
Screening coefficient based  on glmnet coefficients
}
\keyword{internal}
