% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{get_scrcoef}
\alias{get_scrcoef}
\title{Function which generates the screening coefficients}
\usage{
get_scrcoef(scrcoef, data)
}
\arguments{
\item{data}{list of  \code{x} which is a matrix of covariates and
\code{y} which is the response vector}

\item{screeningcoef}{an object of class randomprojection}
}
\description{
Function which generates the screening coefficients
}
\keyword{internal}
