% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{get_screencoef}
\alias{get_screencoef}
\title{Function which generates the screening coefficients}
\usage{
get_screencoef(object, data)
}
\arguments{
\item{object}{an object of class screencoef}

\item{data}{list of  \code{x} which is a matrix of covariates and
\code{y} which is the response vector}
}
\description{
Function which generates the screening coefficients
}
\keyword{internal}
