% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARcv.R
\name{plot.spar.cv}
\alias{plot.spar.cv}
\title{plot.spar.cv}
\usage{
\method{plot}{spar.cv}(
  x,
  plot_type = c("Val_Measure", "Val_numAct", "res-vs-fitted", "coefs"),
  plot_along = c("lambda", "nummod"),
  nummod = NULL,
  lambda = NULL,
  xfit = NULL,
  yfit = NULL,
  opt_par = c("best", "1-se"),
  prange = NULL,
  coef_order = NULL,
  ...
)
}
\arguments{
\item{x}{result of spar.cv function of class "spar.cv".}

\item{plot_type}{one of c("Val_Measure","Val_numAct","res-vs-fitted","coefs").}

\item{plot_along}{one of c("lambda","nummod"); ignored when plot_type="res-vs-fitted".}

\item{nummod}{fixed value for nummod when plot_along="lambda" for plot_type="Val_Measure" or "Val_numAct"; same as for \code{\link{predict.spar.cv}} when plot_type="res-vs-fitted".}

\item{lambda}{fixed value for lambda when plot_along="nummod" for plot_type="Val_Measure" or "Val_numAct"; same as for \code{\link{predict.spar.cv}} when plot_type="res-vs-fitted".}

\item{xfit}{data used for predictions in "res-vs-fitted".}

\item{yfit}{data used for predictions in "res-vs-fitted".}

\item{opt_par}{one of c("best","1se"), only needed for plot_type="res-vs-fitted" to set type of predictions, see \code{\link{predict.spar.cv}}.}

\item{prange}{optional vector of length 2 for "coefs"-plot to give the limits of the predictors' plot range; defaults to c(1,p).}

\item{coef_order}{optional index vector of length p for "coefs"-plot to give the order of the predictors; defaults to 1:p.}

\item{...}{further arguments passed to or from other methods}
}
\value{
ggplot2 object
}
\description{
Plot errors or number of active variables over different thresholds or number of models of spar.cv result, or residuals vs fitted
}
