% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{predict.spar}
\alias{predict.spar}
\title{predict.spar}
\usage{
\method{predict}{spar}(spar_res, xnew, nummod = NULL, lambda = NULL, coef = NULL)
}
\arguments{
\item{spar_res}{result of spar function of class "spar".}

\item{xnew}{matrix of new predictor variables; must have same number of columns as x.}

\item{nummod}{number of models used to form coefficients; value with minimal validation MSE is used if not provided.}

\item{lambda}{threshold level used to form coefficients; value with minimal validation MSE is used if not provided.}

\item{coef}{optional; result of coef.spar, can be used if coef.spar has already been called.}
}
\value{
Vector of predictions
}
\description{
Predict responses for new predictors from spar object
}
