% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARcv.R
\name{print.spar.cv}
\alias{print.spar.cv}
\title{print.spar.cv}
\usage{
\method{print}{spar.cv}(x, ...)
}
\arguments{
\item{x}{result of spar.cv function of class "spar.cv".}

\item{...}{further arguments passed to or from other methods}
}
\value{
text summary
}
\description{
Print summary of spar.cv result
}
