% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{print.spar}
\alias{print.spar}
\title{print.spar}
\usage{
\method{print}{spar}(spar_res)
}
\arguments{
\item{spar_res}{result of spar function of class "spar".}
}
\value{
text summary
}
\description{
Print summary of spar result
}
