% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R
\name{rp_cw}
\alias{rp_cw}
\title{Sparse embedding matrix}
\usage{
rp_cw(..., control = list())
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the random
projection matrix}
}
\value{
object of class randomprojection
}
\description{
The entries of the matrix are generated based on \insertCite{Clarkson2013LowRankApprox}{SPAR}.
}
\details{
Creates an object class randomprojection using arguments passed by user.
}
\references{
{
\insertRef{Clarkson2013LowRankApprox}{SPAR}
}
}
