% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R
\name{rp_cw}
\alias{rp_cw}
\title{Sparse embedding matrix}
\usage{
rp_cw(...)
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the random
projection matrix}
}
\value{
object of class randomprojection
}
\description{
The entries of the matrix are generated based on CW 2013
}
\details{
Creates an object class randomprojection using arguments passed by user.
}
