% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R
\name{rp_gaussian}
\alias{rp_gaussian}
\title{Gaussian random projection matrix}
\usage{
rp_gaussian(...)
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the random
projection matrix}
}
\value{
object of class randomprojection
}
\description{
No arguments need to be passed. The entries of the matrix are generated from
a standard normal distribution.
}
\details{
Creates an object class randomprojection using arguments passed by user.
}
