% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R
\name{rp_sparse}
\alias{rp_sparse}
\title{Sparse random projection matrix}
\usage{
rp_sparse(...)
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the random
projection matrix. The possible argument is \code{psi} in (0,1] which determines
the level of sparsity in the matrix.}
}
\value{
object of class randomprojection
}
\description{
The sparse matrix used in \insertCite{ACHLIOPTAS2003JL}{SPAR} with entries equal to
\eqn{\Psi_{ij} = \pm 1/\sqrt{\psi}} with probability \eqn{\psi/2} and zero otherwise
for \eqn{\psi\in (0,1]}. Default is \code{psi = 1}.
}
\details{
Creates an object class randomprojection using arguments passed by user.
}
\references{
{
\insertRef{ACHLIOPTAS2003JL}{SPAR}
}
}
