% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_classes.R
\name{screen_glmnet}
\alias{screen_glmnet}
\title{Screening coefficient based  on glmnet coefficients}
\usage{
screen_glmnet(..., control = list())
}
\arguments{
\item{...}{includes arguments which can be passed as attributes to the
"\code{screencoef}" object}

\item{control}{list of controls to be passed to the screening function}
}
\value{
object of class "\code{screencoef}" which is a list with elements
\code{name} (character), \code{generate_fun} (function), and \code{control}
(list of controls passed as an argument)
}
\description{
No arguments need to be passed.
}
\details{
Creates an object class "\code{screencoef}" using arguments passed by user.
}
