% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARcv.R
\name{spar.cv}
\alias{spar.cv}
\title{Sparse Projected Averaged Regression}
\usage{
spar.cv(
  x,
  y,
  family = gaussian("identity"),
  model = spar_glmnet(),
  rp = NULL,
  screencoef = NULL,
  nfolds = 10,
  nnu = 20,
  nus = NULL,
  nummods = c(20),
  measure = c("deviance", "mse", "mae", "class", "1-auc"),
  parallel = FALSE,
  seed = NULL,
  set.seed.iteration = FALSE,
  ...
)
}
\arguments{
\item{x}{n x p numeric matrix of predictor variables.}

\item{y}{quantitative response vector of length n.}

\item{family}{a "\code{\link[stats]{family}}" object used for the marginal generalized linear model,
default \code{gaussian("identity")}.}

\item{model}{function creating a "\code{sparmodel}" object; defaults to \code{spar_glmnet()}.}

\item{rp}{function creating a "\code{randomprojection}" object.}

\item{screencoef}{function creating a "\code{screeningcoef}" object}

\item{nfolds}{number of folds to use for cross-validation; should be at least 2, defaults to 10.}

\item{nnu}{number of different threshold values \eqn{\nu} to consider for thresholding;
ignored when \code{nus} is provided; defaults to 20.}

\item{nus}{optional vector of \eqn{\nu}'s to consider for thresholding;
if not provided, \code{nnu} values ranging from 0 to the maximum absolute
marginal coefficient are used.}

\item{nummods}{vector of numbers of marginal models to consider for
validation; defaults to \code{c(20)}.}

\item{measure}{loss to use for validation; defaults to \code{"deviance"}
available for all families. Other options are \code{"mse"} or \code{"mae"}
(between responses and predicted means, for all families),
\code{"class"} (misclassification error) and
\code{"1-auc"} (one minus area under the ROC curve) both just for
binomial family.}

\item{...}{further arguments mainly to ensure back-compatibility}
}
\value{
object of class \code{"spar.cv"} with elements
\itemize{
\item \code{betas} p x  \code{max(nummods)} sparse matrix of class
\code{"\link[=dgCMatrix-class]{dgCMatrix}"} containing the
standardized coefficients from each marginal model
\item \code{intercepts} used in each marginal model, vector of length \code{max(nummods)}
\item \code{scr_coef} p-vector of coefficients used for screening for standardized predictors
\item \code{inds} list of index-vectors corresponding to variables kept after
screening in each marginal model of length  \code{max(nummods)}
\item \code{RPMs} list of projection matrices used in each marginal model of length \code{max(nummods)}
\item \code{val_sum} \code{data.frame} with CV results (mean and sd validation measure and mean number of active variables) for each element of nus and nummods
\item \code{nus} vector of \eqn{\nu}'s considered for thresholding
\item \code{nummods} vector of numbers of marginal models considered for validation
\item \code{ycenter} empirical mean of initial response vector
\item \code{yscale} empirical standard deviation of initial response vector
. \item \code{xcenter} p-vector of empirical means of initial predictor variables
\item \code{xscale} p-vector of empirical standard deviations of initial predictor variables
\item \code{rp} an object of class "\code{randomprojection}"
\item \code{screencoef} an object of class "\code{screeningcoef}"
}
}
\description{
Apply Sparse Projected Averaged Regression to High-dimensional Data (see Parzer, Vana-Guer and Filzmoser 2023).
}
\examples{
\dontrun{
data("example_data")
spar_res <- spar.cv(example_data$x,example_data$y,nummods=c(5,10,15,20,25,30))
spar_res
coefs <- coef(spar_res)
pred <- predict(spar_res,example_data$x)
plot(spar_res)
plot(spar_res,"Val_Meas","nummod")
plot(spar_res,"Val_numAct","nu")
plot(spar_res,"coefs",prange=c(1,400))}
}
\seealso{
\link{spar},\link{coef.spar.cv},\link{predict.spar.cv},\link{plot.spar.cv},\link{print.spar.cv}
}
