% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{spar}
\alias{spar}
\title{Sparse Projected Averaged Regression}
\usage{
spar(
  x,
  y,
  family = gaussian("identity"),
  xval = NULL,
  yval = NULL,
  nlambda = 20,
  lambdas = NULL,
  nummods = c(20),
  type.measure = c("deviance", "mse", "mae", "class", "1-auc"),
  type.rpm = c("cwdatadriven", "cw", "gaussian", "sparse"),
  type.screening = c("ridge", "marglik", "corr"),
  inds = NULL,
  RPMs = NULL,
  control = list(rpm = list(mslow = ceiling(log(ncol(x))), msup = ceiling(nrow(x)/2)),
    scr = list(nscreen = 2 * nrow(x), split_data = FALSE))
)
}
\arguments{
\item{x}{n x p numeric matrix of predictor variables.}

\item{y}{quantitative response vector of length n.}

\item{family}{'family'-objected used for glm (except the quasi), default gaussian("identity").}

\item{xval}{optional matrix of predictor variables observations used for validation of threshold lambda and number of models; x is used if not provided.}

\item{yval}{optional response observations used for validation of threshold lambda and number of models; y is used if not provided.}

\item{nlambda}{number of different lambdas to consider for thresholding; ignored when lambdas are given; defaults to 20.}

\item{lambdas}{optional vector of lambdas to consider for thresholding; if not provided, nlam values ranging from 0 to the maximum ablsolute marginal coefficient are used.}

\item{nummods}{vector of numbers of marginal models to consider for validation; defaults to c(20).}

\item{type.measure}{loss to use for validation; defaults to "deviance" available for all families. Other options are "mse" or "mae" (between responses and predicted means, for all families),
"class" (misclassification error) and "1-auc" (one minus area under the ROC curve) both just for "binomial" family.}

\item{type.rpm}{type of random projection matrix to be employed; one of "cwdatadriven", "cw", "gaussian", "sparse"; defaults to "cwdatadriven".}

\item{type.screening}{type of screening coefficients; one of "ridge", "marglik", "corr"; defaults to "ridge" which is based on the ridge coefficients where the penalty converges to zero.}

\item{inds}{optional list of index-vectors corresponding to variables kept after screening in each marginal model of length max(nummods),dimensions need to fit those of RPMs.}

\item{RPMs}{optional list of sparse CW projection matrices used in each marginal model of length max(nummods), diagonal elements will be overwritten with a coefficient only depending on the given x and y.}

\item{control}{a list two elements: rpm and scr. rpm contains a list of optional arguments to be passed to functions creating the random projection matrices. Here mslow is a lower bound for uniform random goal dimensions in marginal models; defaults to log(p);
msup is upper bound for uniform random goal dimensions in marginal models; defaults to n/2.
scr contains a list of optional arguments to be passed to functions performing screening. nscreen is the number of variables to keep after screening 2n;
split_data logical to indicate whether data for calculation of scr_coef and fitting of mar mods should be split 1/4 to 3/4 to avoid overfitting; default FALSE.}
}
\value{
object of class "spar" with elements
\itemize{
\item betas p x max(nummods) matrix of standardized coefficients from each marginal model
\item scr_coef p-vector of HOLP coefficient used for screening
\item inds list of index-vectors corresponding to variables kept after screening in each marginal model of length max(nummods)
\item RPMs list of sparse CW projection matrices used in each marginal model of length max(nummods)
\item val_res data.frame with validation results (validation measure and number of active variables) for each element of lambdas and nummods
\item val_set logical flag, whether validation data were provided; if FALSE, training data were used for validation
\item lambdas vector of lambdas considered for thresholding
\item nummods vector of numbers of marginal models considered for validation
\item xcenter p-vector of empirical means of initial predictor variables
\item xscale p-vector of empirical standard deviations of initial predictor variables
}
}
\description{
Apply Sparse Projected Averaged Regression to High-dimensional Data (see Parzer, Vana-Guer and Filzmoser 2023).
This function performs the procedure for given thresholds lambda and numbers of marginal models, and acts as a help-function for the full cross-validated procedure \link{spar.cv}.
}
\examples{
\dontrun{
data("example_data")
spar_res <- spar(example_data$x,example_data$y,
xval=example_data$xtest,yval=example_data$ytest,nummods=c(5,10,15,20,25,30))
spar_res
coefs <- coef(spar_res)
pred <- predict(spar_res,xnew=example_data$x)
plot(spar_res)
plot(spar_res,"Val_Meas","nummod")
plot(spar_res,"Val_numAct","lambda")
plot(spar_res,"coefs",prange=c(1,400))}
}
\seealso{
\link{spar.cv},\link{coef.spar},\link{predict.spar},\link{plot.spar},\link{print.spar}
}
