% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAR.R
\name{spar}
\alias{spar}
\title{Sparse Projected Averaged Regression}
\usage{
spar(
  x,
  y,
  family = gaussian("identity"),
  xval = NULL,
  yval = NULL,
  nnu = 20,
  nus = NULL,
  nummods = c(20),
  type.measure = c("deviance", "mse", "mae", "class", "1-auc"),
  type.rpm = c("cwdatadriven", "cw", "gaussian", "sparse"),
  type.screening = c("ridge", "marglik", "corr"),
  inds = NULL,
  RPMs = NULL,
  rp = NULL,
  scrcoef = NULL,
  control = list(rpm = list(mslow = ceiling(log(ncol(x))), msup = ceiling(nrow(x)/2)),
    scr = list(nscreen = 2 * nrow(x), split_data = FALSE))
)
}
\arguments{
\item{x}{n x p numeric matrix of predictor variables.}

\item{y}{quantitative response vector of length n.}

\item{family}{a \code{\link[stats]{"family"}} object used for the marginal generalized linear model,
default \code{gaussian("identity")}.}

\item{xval}{optional matrix of predictor variables observations used for
validation of threshold nu and number of models; \code{x} is used
if not provided.}

\item{yval}{optional response observations used for validation of
threshold nu and number of models; \code{y} is used if not provided.}

\item{nnu}{number of different threshold values \eqn{\nu} to consider for thresholding;
ignored when nus are given; defaults to 20.}

\item{nus}{optional vector of \eqn{\nu}'s to consider for thresholding;
if not provided, \code{nnu} values ranging from 0 to the maximum absolute
marginal coefficient are used.}

\item{nummods}{vector of numbers of marginal models to consider for
validation; defaults to \code{c(20)}.}

\item{type.measure}{loss to use for validation; defaults to \code{"deviance"}
available for all families. Other options are \code{"mse"} or \code{"mae"}
(between responses and predicted means, for all families),
\code{"class"} (misclassification error) and
\code{"1-auc"} (one minus area under the ROC curve) both just for
binomial family.}

\item{type.rpm}{type of random projection matrix to be employed;
one of \code{"cwdatadriven"},
\code{"cw"} \insertCite{Clarkson2013LowRankApprox}{SPAR},
\code{"gaussian"}, \code{"sparse"} \insertCite{ACHLIOPTAS2003JL}{SPAR};
defaults to \code{"cwdatadriven"}.}

\item{type.screening}{type of screening coefficients; one of \code{"ridge"},
\code{"marglik"}, \code{"corr"}; defaults to \code{"ridge"} which is
based on the ridge coefficients where the penalty converges to zero.}

\item{inds}{optional list of index-vectors corresponding to variables kept
after screening in each marginal model of length \code{max(nummods)};
dimensions need to fit those of RPMs.}

\item{RPMs}{optional list of projection matrices used in each
marginal model of length \code{max(nummods)}, diagonal elements will be
overwritten with a coefficient only depending on the given \code{x} and \code{y}.}

\item{control}{a list two elements: \code{rpm} and \code{scr}.
Element \code{rpm} contains a list of optional arguments to be passed to
functions creating the random projection matrices. Here \code{mslow} is a
lower bound for uniform random goal dimensions in
marginal models; defaults to \eqn{\log(p)};
\code{msup} is upper bound for uniform random goal dimensions in marginal models;
defaults to n/2.
Element \code{scr} contains a list of optional arguments to be passed to
functions performing screening: \code{nscreen} is the number of variables to keep after screening 2n;
\code{split_data} logical to indicate whether data for calculation of screening coefficient
and fitting of mar mods should be split 1/4 to 3/4 to avoid overfitting;
default \code{FALSE}.}
}
\value{
object of class \code{"spar"} with elements
\itemize{
\item betas p x \code{max(nummods)} matrix of standardized coefficients from each
marginal model
\item scr_coef p-vector of coefficients used for screening for standardized predictors
\item inds list of index-vectors corresponding to variables kept after screening in each marginal model of length max(nummods)
\item RPMs list of projection matrices used in each marginal model of length \code{max(nummods)}
\item val_res \code{data.frame} with validation results (validation measure
and number of active variables) for each element of \code{nus} and \code{nummods}
\item val_set logical flag, whether validation data were provided;
if \code{FALSE}, training data were used for validation
\item nus vector of \eqn{\nu}'s considered for thresholding
\item nummods vector of numbers of marginal models considered for validation
\item xcenter p-vector of empirical means of initial predictor variables
\item xscale p-vector of empirical standard deviations of initial predictor variables
}
}
\description{
Apply Sparse Projected Averaged Regression to High-dimensional Data
\insertCite{parzer2024lm}{SPAR} \insertCite{parzer2024glms}{SPAR}.
This function performs the procedure for given thresholds nu and numbers
of marginal models, and acts as a help-function for the full cross-validated
procedure \link{spar.cv}.
}
\examples{
\dontrun{
data("example_data")
spar_res <- spar(example_data$x,example_data$y,
xval=example_data$xtest,yval=example_data$ytest,nummods=c(5,10,15,20,25,30))
spar_res
coefs <- coef(spar_res)
pred <- predict(spar_res,xnew=example_data$x)
plot(spar_res)
plot(spar_res,"Val_Meas","nummod")
plot(spar_res,"Val_numAct","nu")
plot(spar_res,"coefs",prange=c(1,400))}
}
\references{
{
\insertRef{parzer2024lm}{SPAR}

\insertRef{parzer2024glms}{SPAR}

\insertRef{Clarkson2013LowRankApprox}{SPAR}

\insertRef{ACHLIOPTAS2003JL}{SPAR}
}
}
\seealso{
\link{spar.cv},\link{coef.spar},\link{predict.spar},\link{plot.spar},\link{print.spar}
}
